Refactor #84: Unify package/module visibility rules and clean up the public API.
- A subpackage is **public** iff its name is **unprefixed**.
- A module is **private** iff it has a `_` prefix.
- These conditions are **mutually exclusive** (never both).

## Changes
- `tabicl.{model,sklearn,unsupervised}` → `_model`, `_sklearn`, `_unsupervised`
- Internal files inside private packages drop their `_` prefix (privacy marked once at package level).
- `tabicl.forecast.transforms` remains public (for `TimeTransform` subclassing), but its implementation files gain a `_` prefix.

## API Changes
- `TabICL` and `TabICLCache` are **no longer user-facing**.
- Users should migrate to `TabICLClassifier` / `TabICLRegressor` / `TabICLForecaster`.
- Public top-level imports remain unchanged.

## Bug Fixes
- Fixed latent `from tabicl import TabICL` import bug in `train/_run.py`.

## Documentation
- Added docstrings clarifying boundaries for `tabicl.prior` and `tabicl.train`.
- Updated `docs/api.rst` to exactly match `__all__`; removed docs for private `tabicl.model.TabICL`.
- Added "In development" entry to `CHANGES.md`.
