+++ src/tabicl/_model/attention.py
@@ -6,8 +6,8 @@ import torch
 from torch import Tensor, nn
 from torch.nn import functional as F
 
-from ._rope import RotaryEmbedding
-from ._kv_cache import KVCacheEntry
+from .rope import RotaryEmbedding
+from .kv_cache import KVCacheEntry
 
 try:
     from flash_attn_interface import flash_attn_varlen_func as flash_attn3
