+++ src/tabicl/_model/embedding.py
@@ -8,11 +8,11 @@ import torch
 from torch import nn, Tensor
 import torch.nn.functional as F
 
-from ._layers import SkippableLinear, OneHotAndLinear
-from ._encoders import SetTransformer
-from ._kv_cache import KVCache
-from ._inference import InferenceManager
-from ._inference_config import MgrConfig, InferenceConfig
+from .layers import SkippableLinear, OneHotAndLinear
+from .encoders import SetTransformer
+from .kv_cache import KVCache
+from .inference import InferenceManager
+from .inference_config import MgrConfig, InferenceConfig
 
 
 class ColEmbedding(nn.Module):
