+++ src/tabicl/_model/inference_config.py
