+++ src/tabicl/_model/quantile_dist.py
