+++ src/tabicl/_model/ssmax.py
