+++ src/tabicl/_sklearn/__init__.py
@@ -1,0 +1,4 @@
+"""Private. Access estimators via ``from tabicl import TabICLClassifier, TabICLRegressor``."""
+
+from .classifier import TabICLClassifier
+from .regressor import TabICLRegressor
