+++ src/tabicl/_unsupervised/__init__.py
@@ -1,0 +1,3 @@
+"""Private. Access TabICLUnsupervised via ``from tabicl import TabICLUnsupervised``."""
+
+from .unsupervised import TabICLUnsupervised
