+++ src/tabicl/forecast/transforms/_pipeline.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import pandas as pd
 
 from tabicl.forecast._ts_dataframe import TimeSeriesDataFrame
-from tabicl.forecast.transforms.base import TimeTransform
+from tabicl.forecast.transforms._base import TimeTransform
 
 
 class TimeTransformChain:
