--- src/tabicl/model/__init__.py
@@ -1,5 +1,0 @@
-from ._tabicl import TabICL
-from ._kv_cache import TabICLCache
-from ._inference_config import InferenceConfig
-
-__all__ = ["TabICL", "InferenceConfig"]
