--- src/tabicl/sklearn/__init__.py
@@ -1,4 +1,0 @@
-from ._classifier import TabICLClassifier
-from ._regressor import TabICLRegressor
-
-__all__ = ["TabICLClassifier", "TabICLRegressor"]
