+++ src/tabicl/train/__init__.py
@@ -1,1 +1,8 @@
-__all__ = []  # Internal subpackage. Use `python -m tabicl.train` as the CLI entry point.
+"""CLI entry point subpackage.
+
+This subpackage exists only to support ``python -m tabicl.train``. Importing
+from ``tabicl.train.*`` in Python code is NOT public API and may change
+without notice.
+"""
+
+__all__ = []
